Updated build system
- Factored out require statements in gulpfile.coffee
- Move transforms into package.json in this project for consistency with
gavel which had transforms moved into package.json so that the package
would be rebuilt correctly when it changed (see the browserify handbook
section about browserify.transform field in package.json)
- Extract sourcemaps to external file for performance and because there
seems to be a bug that prevents rebuilds from working correctly
otherwise
- Add babelify back into the mix. I'd like babelify in production and it
doesn't seem to be affecting build times all that much in development so
it's a bit easier and cleaner to have the same transforms for production
and development builds.
- Remove coffee tags. The current plan is to use jsctags with source
maps which should give us the full power of Tern JS for navigating.
- Update the package.lock file