Break MapController out into various strategies for different phases.
- Allows us to implement a controller strategy per phase to keep the
MapController from becoming bloated and hopefully make it easier to
isolate and navigate the controls for a particular phase.

- Map initialization and determining which set of controls to use has
been left up to the MapController. Any other common logic should go
there as well.

- Create KeyboardInputHandler class to deal with interpreting the
meaning and state of various keystrokes. It's an easy way to make the
controller strategies and the map controller a bit thinner and it makes
logical sense to separate this out.